refactor(error): separate error module
